requirements.txt
asgiref==3.8.1
# black==24.8.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
Django==5.0.6
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-filter==24.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==25.2.0
# flake8==7.1.1
idna==3.7
iniconfig==2.0.0
mccabe==0.7.0
# mypy==1.11.1
# mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
# pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
stripe==11.1.0
